+++ Presentation/DTOs/PurchasedProductDTO.cs
@@ -15,7 +15,7 @@ namespace HyHeroesWebAPI.Presentation.DTOs
 
         [Required]
         [JsonProperty("pricePerMonth")]
-        public decimal PricePerMonth { get; set; }
+        public int PricePerMonth { get; set; }
 
         [Required]
         [JsonProperty("imageUrl")]
@@ -23,7 +23,7 @@ namespace HyHeroesWebAPI.Presentation.DTOs
 
         [Required]
         [JsonProperty("permanentPrice")]
-        public decimal PermanentPrice { get; set; }
+        public int PermanentPrice { get; set; }
 
 #nullable enable
         [JsonProperty("description")]
